Added new map center
