+++ pages/Map.jsx
@@ -1,0 +1,28 @@
+import { useState, useEffect } from "react";
+import "leaflet/dist/leaflet.css";
+import { MapContainer, TileLayer, Marker, useMap } from "react-leaflet";
+
+export function ChangeView({ coords }) {
+  const map = useMap();
+  map.setView(coords, 12);
+  return null;
+}
+
+export default function Map() {
+  const [geoData, setGeoData] = useState({ lat: -8.7445, lng: 115.182 });
+
+  const center = [geoData.lat, geoData.lng];
+
+  return (
+    <MapContainer
+      center={center}
+      zoom={12}
+      style={{ height: "50%", width: "100%" }}
+    >
+      <TileLayer
+        attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
+        url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
+      />
+    </MapContainer>
+  );
+}
