+++ pages/index.tsx
@@ -17,9 +17,8 @@ export default function Home() {
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <main className={styles.main}>  
-      <h1> Hello Team</h1>
+      <h1>Hello Team</h1>
       <MapWithNoSSR />
-    
       </main>
     </>
   )
